requirements.txt
# Flask
Flask>=3.0.0
Flask-Session>=0.8.0
gunicorn>=22.0.0

# Session store
redis>=5.0.0

# Auth
msal>=1.28.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
requests>=2.32.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.3.0

# Encryption
cryptography>=42.0.0

# AI / LLM
anthropic>=0.40.0
openai>=1.30.0
voyageai>=0.3.0
tiktoken>=0.7.0

# Import (Excel support)
openpyxl>=3.1.0

# External DB sync (location hierarchy)
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Billing
stripe>=10.0.0

# Phone helpdesk (ElevenLabs + Twilio)
twilio>=9.0.0

# Email
resend>=2.0.0

# Scraping (KB scrapers)
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0
playwright>=1.40.0

# Logging
python-json-logger>=2.0.0

# Rate limiting
Flask-Limiter[redis]>=3.5.0

# Observability
sentry-sdk[flask]>=2.0.0

# Document parsing (KB article uploads)
PyPDF2>=3.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
